LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/pseudo-elements-003.tentative-expected.txt

First-line
First-line

PASS Originating element container for ::before
PASS Originating element container for ::after
PASS Originating element container for ::marker
PASS Originating element container for ::first-line
PASS Originating element container for ::first-letter
PASS Originating element container for outer ::first-line
PASS Originating element container for outer ::first-letter
FAIL Originating element container for ::backdrop assert_equals: ::backdrop not rendered expected "rgb(0, 255, 0)" but got "rgb(0, 128, 0)"
